+++ extractfuncs/l_script.c
@@ -813,7 +813,7 @@ int PS_ReadPunctuation( script_t *script, token_t *token ) {
 		punc = &script->punctuations[i];
 #endif //PUNCTABLE
 		p = punc->p;
-		len = strlen( p );
+		len = (int)strlen( p );
 		//if the script contains at least as much characters as the punctuation
 		if ( script->script_p + len <= script->end_p ) {
 			//if the script contains the punctuation
@@ -1274,7 +1274,7 @@ int ScriptSkipTo( script_t *script, char *value ) {
 	char firstchar;
 
 	firstchar = *value;
-	len = strlen( value );
+	len = (int)strlen( value );
 	do
 	{
 		if ( !PS_ReadWhiteSpace( script ) ) {
